+++ point3D.h
@@ -3,23 +3,23 @@
 
 class Point3D {
     public:
-        Point3D(float x, float y, float z);
+        Point3D(double x, double y, double z);
         ~Point3D();
         Point3D operator+(const Point3D &rhs);
         Point3D operator-(const Point3D &rhs);
-        Point3D operator*(const float &rhs);
-        Point3D operator/(const float &rhs);
+        Point3D operator*(const double &rhs);
+        Point3D operator/(const double &rhs);
         friend std::ostream& operator<<(std::ostream& os, Point3D const &point){
             os << "{" << point.m_x << ", " << point.m_y << ", " << point.m_z << "}";
             return os;
         };
-        float getX();
-        float getY();
-        float getZ();
-        float norm();
+        double getX();
+        double getY();
+        double getZ();
+        double norm();
     private:
-        float m_x;
-        float m_y;
-        float m_z;
+        double m_x;
+        double m_y;
+        double m_z;
 
 };
