+++ vectorMaths.h
@@ -6,7 +6,7 @@ class VectorMaths {
         VectorMaths();
         ~VectorMaths();
 
-        float dot(Point3D p1 , Point3D p2);
+        double dot(Point3D p1 , Point3D p2);
         Point3D cross(Point3D p1 , Point3D p2);
         void normalise(Point3D* p1);
     private:
